requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.7.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.3.4
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
stripe==7.8.0
